Add binConverter and BinEdit
Binedit tries to edit a file within the console.
binConverter tries to change the endianness of a file that the segments of the data are 4 bytes long. if the last bytes do not match 4 bytes then it will throw a like - out of bound error. This was meant to try to change the content of N64 files: .n64 z64, v64